+++ benchmark/data_structure/atomic.cc
@@ -187,7 +187,7 @@ class AtomicInstructionTestSuite {
  * */
 template<typename AtomicT, typename T, std::memory_order MemoryOrder>
 void TestAtomicInstructionsPerThread() {
-  int count = (1<<20);
+  size_t count = (1<<20);
   AtomicInstructionTestSuite<AtomicT, T, MemoryOrder>().AtomicIncrement(count);
   AtomicInstructionTestSuite<AtomicT, T, MemoryOrder>().AtomicFetchAdd(count);
   AtomicInstructionTestSuite<AtomicT, T, MemoryOrder>().AtomicAssign(count);
